Refactor and fix sharedwith cache root
